Add files via upload
My project is about Progressive Overload. It is an app that is intended for an audience that goes to the gym regularly and lifts weights. Progressive Overload is an important aspect when it comes to lifting weights and people usually do not have a good way to track it - It is disorganized which is why I wish to have an app made, for it to be feasibly tracked. There will be a selection of exercises in the app which the users themselves add. Then they enter the "reps", "sets" and "weight" after the workout for the exercise that was selected. There should be a view page in the app where the progress of each workout is written and shown in a graphical format. Based on this, the user can track their progressive overload.